feat: Add function to calculate sum of two numbers
Add a new function `add` to calculate the sum of two numbers. This function is implemented in the `polars-use` library. A test case is also added to verify the correctness of the `add` function.

chore: Update dependencies in Cargo.toml

Update the dependencies in the `Cargo.toml` file of the root project. The `anyhow` dependency is updated to version 1.0.91 and the `thiserror` dependency is updated to version 1.0.65. The `polars` dependency is updated to version 0.43.1 and additional features are enabled.

feat: Add polars-use to workspace members

Add the `polars-use` library to the workspace members in the `Cargo.toml` file of the root project.

feat: Add dependencies to polars-use Cargo.toml

Add the necessary dependencies to the `Cargo.toml` file of the `polars-use` library. The `anyhow`, `chrono`, and `polars` dependencies are added, along with the `utils` dependency from a local path.

chore: Add polars-use source file

Add the source file `lib.rs` to the `polars-use` library. This file contains the implementation of the `add` function and a test case.